Some insight on the DxFlags
The Int32 at 0x1c acts more as a "usage" flag from what I have been seeing. There are parts in the code that flip the endian of every field in the FTEX header and this region seems to be swapped as a full 32bit integer rather than two distinct shorts.